// src/shared/enum/end-points/app.enum.ts
export enum AppEndPointsEnum {
  PREFIX = 'api',
  USER = 'user',
  USER_ADMIN = 'user/admin',
  AUTH = 'auth',
  NEWS = 'news',
  NEWS_ADMIN = 'news/admin',
  BRAND_SUPER_ADMIN = 'brand/admin',
  BRAND_PUBLIC = 'brand',
  BRAND_BRAND_ADNIN = 'brand/brandAdmin',
  WALLET = 'wallet',
  EDUCATION_SUPER_ADMIN = 'education/admin',
  VIDEO_SUPER_ADMIN = 'video/admin',
  UPLOAD = 'upload',
  VIDEO_BRAND_ADMIN = 'video/brandAdmin',
  VIDEO_PUBLIC = 'video',
  EDUCATION_Brand_ADMIN = 'education/brandAdmin',
  EDUCATION_PUBLIC = 'education',
  COMMENT = 'comment/:type',
  COMMENT_SUPER_ADMIN = 'comment/admin',
  COMMENT_BRAND_ADMIN = 'comment/education/brandAdmin',
  EVENT_SUPER_ADMIN = 'event/admin',
  EVENT_BRAND_ADMIN = 'event/brandAdmin',
  EVENT_PUBLIC = 'event',
  CART_USER = 'cart',
  PAYMETN = '/payment',
  DISCOUNT_SUPER_ADMIN = '/discount/admin',
  DISCOUNT = '/discount',
}