// src/user/enum/message.enum.ts
export enum UserMessageEnum {
  EXIST_USERNAME = 'نام‌کاربری تکراری است',
  EXIST_EMAIL = 'ایمیل تکراری است',
  EXIST_CELL_NUMBER = 'شماره همراه تکراری است',
  NOT_FOUND_USER = 'کاربر پیدا نشد',
  INVALID_PASSWORD = 'رمز عبور صحیح نیست',
  CREATE_USER_BY_ADMIN = 'create new user By ',
  DELETE_USER_BY_ADMIN = 'delete user By ',
  BAN_USER_BY_ADMIN = 'ban user By ',
  UNBAN_USER_BY_ADMIN = 'unban new user By ',
  ADD_NEW_ADMIN = 'Adding new admin by ',
  REMOVE_ADMIN = 'Remove a admin by ',
  SUCCESS_BRAND_ADMIN = 'کاربر با موفقیت بن شد',
}